+++ src/lib/coil/vxworks/coil/OS.cpp
@@ -1,0 +1,19 @@
+﻿// -*- C++ -*-
+/*!
+ * @file OS_vxworks.cpp
+ * @brief OS class
+ * @date $Date$
+ * @author Nobuhiko Miyamoto<n-miyamoto@aist.go.jp>
+ *
+ * Copyright (C) 2019
+ *     Robot Innovation Research Center,
+ *     Intelligent Systems Research Institute,
+ *     National Institute of
+ *         Advanced Industrial Science and Technology (AIST), Japan
+ *     All rights reserved.
+ *
+ * $Id$
+ *
+ */
+
+#include <coil/OS.h>
