+++ src/lib/coil/win32/coil/Mutex.cpp
@@ -18,3 +18,24 @@
  */
 
 #include <coil/Mutex.h>
+
+namespace coil
+{
+    Mutex::Mutex(const char * name)
+    {
+      SECURITY_DESCRIPTOR sd_buffer;
+      ::InitializeSecurityDescriptor(&sd_buffer,
+                                     SECURITY_DESCRIPTOR_REVISION);
+      ::SetSecurityDescriptorDacl(&sd_buffer, TRUE, 0, FALSE);
+      m_Security_attr.nLength = sizeof(SECURITY_ATTRIBUTES);
+      m_Security_attr.lpSecurityDescriptor = &sd_buffer;
+      m_Security_attr.bInheritHandle = TRUE;
+      mutex_ = ::CreateMutexA(&m_Security_attr, FALSE, name);
+    }
+
+    Mutex:: ~Mutex()
+    {
+      ::CloseHandle(mutex_);
+    }
+
+}
