+++ src/lib/coil/win32/coil/Mutex.h
@@ -61,19 +61,7 @@ namespace coil
      *
      * @endif
      */
-    explicit Mutex(const char * const name = 0)
-    {
-      SECURITY_DESCRIPTOR sd_buffer;
-      ::InitializeSecurityDescriptor(&sd_buffer,
-                                     SECURITY_DESCRIPTOR_REVISION);
-      ::SetSecurityDescriptorDacl(&sd_buffer, TRUE, 0, FALSE);
-      m_Security_attr.nLength = sizeof(SECURITY_ATTRIBUTES);
-      m_Security_attr.lpSecurityDescriptor = &sd_buffer;
-      m_Security_attr.bInheritHandle = TRUE;
-      mutex_ = ::CreateMutexA(&m_Security_attr, FALSE, name);
-
-
-    }
+    explicit Mutex(const char * const name = 0);
 
     /*!
      * @if jp
@@ -90,10 +78,7 @@ namespace coil
      *
      * @endif
      */
-    ~Mutex()
-    {
-      ::CloseHandle(mutex_);
-    }
+    ~Mutex();
 
     /*!
      * @if jp
