+++ src/lib/coil/win32/coil/OS.h
@@ -51,17 +51,10 @@ namespace coil
   class osversion
   {
   public:
-      osversion() : major(0), minor(0)
-      {
+      osversion();
+      osversion(const osversion& obj);
 
-      }
-      osversion(const osversion& obj) : major(obj.major), minor(obj.minor)
-      {
-
-      }
-      osversion(DWORD majar_version,DWORD minor_version) : major(majar_version), minor(minor_version)
-      {
-      }
+      osversion(DWORD majar_version,DWORD minor_version);
       DWORD major;
       DWORD minor;
   };
@@ -92,45 +85,7 @@ namespace coil
   *
   * @endif
   */
-  inline bool rtlgetinfo(utsname* name)
-  {
-      HMODULE handle = GetModuleHandle("ntdll.dll");
-
-      if (handle)
-      {
-          typedef LONG(WINAPI* RtlGetVersionFunc)(PRTL_OSVERSIONINFOW lpVersionInformation);
-
-          RTL_OSVERSIONINFOW version_info;
-          RtlGetVersionFunc RtlGetVersion = (RtlGetVersionFunc)GetProcAddress(handle, "RtlGetVersion");
-          if (RtlGetVersion != nullptr)
-          {
-              if (RtlGetVersion(&version_info) == 0)
-              {
-                  const char *os;
-                  if (version_info.dwPlatformId == VER_PLATFORM_WIN32_NT)
-                  {
-                      os = "Windows NT %d.%d";
-                  }
-                  else
-                  {
-                      os = "Windows CE %d.%d";
-                  }
-
-                  sprintf_s(name->release, sizeof(name->release), os,
-                      static_cast<int>(version_info.dwMajorVersion),
-                      static_cast<int>(version_info.dwMinorVersion));
-
-                  sprintf_s(name->version, sizeof(name->release), "Build %d %ls",
-                      static_cast<int>(version_info.dwBuildNumber),
-                      version_info.szCSDVersion);
-
-                  return true;
-              }
-          }
-          FreeLibrary(handle);
-      }
-      return false;
-  }
+  bool rtlgetinfo(utsname* name);
 
   /*!
    * @if jp
@@ -168,172 +123,7 @@ namespace coil
    *
    * @endif
    */
-  inline int uname(utsname* name)
-  {
-    if (name == NULL) return 0;
-    int ret(0);
-
-    // name.sysname
-    ::strcpy_s(name->sysname, sizeof(name->sysname), "Win32");
-
-
-    
-    if(!rtlgetinfo(name))
-    {
-
-        std::map<std::string, osversion> oslist;
-        oslist["Windows 2000"] = osversion(5, 0);
-        oslist["Windows XP 32bit"] = osversion(5, 1);
-        oslist["Windows XP 64bit"] = osversion(5, 2);
-        oslist["Windows Vista"] = osversion(6, 0);
-        oslist["Windows 7"] = osversion(6, 1);
-        oslist["Windows 8"] = osversion(6, 2);
-        oslist["Windows 8.1"] = osversion(6, 3);
-        oslist["Windows 10"] = osversion(10, 0);
-
-        for(auto & o : oslist)
-        {
-            // name.release, name.version
-            OSVERSIONINFOEX version_info;
-            ULONGLONG condition = 0;
-            version_info.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
-            version_info.dwMajorVersion = o.second.major;
-            version_info.dwMinorVersion = o.second.minor;
-            VER_SET_CONDITION(condition, VER_MAJORVERSION, VER_EQUAL);
-            VER_SET_CONDITION(condition, VER_MINORVERSION, VER_EQUAL);
-
-            if (::VerifyVersionInfo(&version_info, VER_MAJORVERSION | VER_MINORVERSION, condition))
-            {
-                const char *os;
-                if (version_info.dwPlatformId == VER_PLATFORM_WIN32_NT)
-                {
-                    os = "Windows NT %d.%d";
-                }
-                else
-                {
-                    os = "Windows CE %d.%d";
-                }
-
-                sprintf_s(name->release, sizeof(name->release), os,
-                    static_cast<int>(version_info.dwMajorVersion),
-                    static_cast<int>(version_info.dwMinorVersion));
-
-                break;
-            }
-
-
-        }
-    }
-
-
-    // name.machine
-    SYSTEM_INFO sys_info;
-    GetSystemInfo(&sys_info);
-    WORD arch = sys_info.wProcessorArchitecture;
-    char cputype[COIL_UTSNAME_LENGTH/2];
-    char subtype[COIL_UTSNAME_LENGTH/2];
-
-    switch (arch)
-      {
-      case PROCESSOR_ARCHITECTURE_INTEL:
-        strcpy_s(cputype, sizeof(cputype), "x86 Intel");
-        if (sys_info.wProcessorLevel == 3)
-          strcpy_s(subtype, sizeof(subtype), "80386");
-        else if (sys_info.wProcessorLevel == 4)
-          strcpy_s(subtype, sizeof(subtype), "80486");
-        else if (sys_info.wProcessorLevel == 5)
-          strcpy_s(subtype, sizeof(subtype), "Pentium");
-        else if (sys_info.wProcessorLevel == 6)
-          strcpy_s(subtype, sizeof(subtype), "Pentium Pro");
-        else if (sys_info.wProcessorLevel == 7)
-          strcpy_s(subtype, sizeof(subtype), "Pentium II");
-        else
-          strcpy_s(subtype, sizeof(subtype), "Pentium Family");
-        break;
-      case PROCESSOR_ARCHITECTURE_MIPS:
-          strcpy_s(cputype, sizeof(cputype), "MIPS");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_ALPHA:
-          strcpy_s(cputype, sizeof(cputype), "ALPHA");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_PPC:
-          strcpy_s(cputype, sizeof(cputype), "PPC");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_SHX:
-          strcpy_s(cputype, sizeof(cputype), "SHX");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_ARM:
-          strcpy_s(cputype, sizeof(cputype), "ARM");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_IA64:
-          strcpy_s(cputype, sizeof(cputype), "IA64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_ALPHA64:
-          strcpy_s(cputype, sizeof(cputype), "ALPHA64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_MSIL:
-          strcpy_s(cputype, sizeof(cputype), "MSIL");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_AMD64:
-          if (sys_info.wProcessorLevel == 6)
-          {
-              strcpy_s(cputype, sizeof(cputype), "x64 Intel");
-              strcpy_s(subtype, sizeof(subtype), "Pentium Pro");
-          }
-          else
-          {
-              strcpy_s(cputype, sizeof(cputype), "x64(AMD or Intel)");
-              strcpy_s(subtype, sizeof(subtype), "");
-          }
-          break;
-      case PROCESSOR_ARCHITECTURE_IA32_ON_WIN64:
-          strcpy_s(cputype, sizeof(cputype), "IA32_ON_WIN64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-      case PROCESSOR_ARCHITECTURE_NEUTRAL:
-          strcpy_s(cputype, sizeof(cputype), "NEUTRAL");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-#ifdef PROCESSOR_ARCHITECTURE_ARM64
-      case PROCESSOR_ARCHITECTURE_ARM64:
-          strcpy_s(cputype, sizeof(cputype), "ARM64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-#endif
-#ifdef PROCESSOR_ARCHITECTURE_ARM32_ON_WIN64
-      case PROCESSOR_ARCHITECTURE_ARM32_ON_WIN64:
-          strcpy_s(cputype, sizeof(cputype), "ARM32_ON_WIN64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-#endif
-#ifdef PROCESSOR_ARCHITECTURE_IA32_ON_ARM64
-      case PROCESSOR_ARCHITECTURE_IA32_ON_ARM64:
-          strcpy_s(cputype, sizeof(cputype), "IA32_ON_ARM64");
-          strcpy_s(subtype, sizeof(subtype), "");
-          break;
-#endif
-      default:
-        strcpy_s(cputype, sizeof(cputype), "Unknown");
-        strcpy_s(subtype, sizeof(subtype), "Unknown");
-      }
-    sprintf_s(name->machine, sizeof(name->machine), "%s %s", cputype, subtype);
-
-    // name.nodename
-    DWORD len = COIL_UTSNAME_LENGTH;
-    if (GetComputerNameExA(ComputerNameDnsHostname,
-                            name->nodename,
-                            &len) == false)
-    ret = -1;
-    return ret;
-  }
+  int uname(utsname* name);
 
   /*!
    * @if jp
@@ -355,10 +145,7 @@ namespace coil
    * @endif
    */
   typedef int pid_t;
-  inline pid_t getpid()
-  {
-    return ::_getpid();
-  }
+  pid_t getpid();
 
   /*!
    * @if jp
@@ -379,10 +166,7 @@ namespace coil
    *
    * @endif
    */
-  inline pid_t getppid()
-  {
-    return 0;
-  }
+  pid_t getppid();
 
   /*!
    * @if jp
@@ -407,18 +191,7 @@ namespace coil
    *
    * @endif
    */
-  inline char* getenv(const char* name)
-  {
-    size_t return_size;
-    ::getenv_s(&return_size, NULL, 0, name);
-    if (return_size == 0)
-    {
-        return nullptr;
-    }
-    char *buff = new char[return_size * sizeof(char)];
-    ::getenv_s(&return_size, buff, return_size, name);
-    return buff;
-  }
+  bool getenv(const char* name, std::string &env);
 
   static int    opterr = 1,     /* if error message should be printed */
                 optind = 1,     /* index into parent argv vector */
@@ -595,13 +368,7 @@ namespace coil
      *
      * @endif
      */
-    GetOpt(int argc, char* const argv[], const char* opt, int flag)
-      : optind(1), opterr(1), optopt(0),
-        m_argc(argc), m_argv(argv), m_opt(opt), m_flag(flag)
-    {
-      this->optarg = coil::optarg;
-      coil::optind = 1;
-    }
+    GetOpt(int argc, char* const argv[], const char* opt, int flag);
 
     /*!
      * @if jp
@@ -618,10 +385,7 @@ namespace coil
      *
      * @endif
      */
-    ~GetOpt()
-    {
-      coil::optind = 1;
-    }
+    ~GetOpt();
 
     /*!
      * @if jp
@@ -642,19 +406,7 @@ namespace coil
      *
      * @endif
      */
-    int operator()()
-    {
-      coil::opterr = opterr;
-      coil::optind = optind;
-
-      int result = getopt(m_argc, m_argv, m_opt);
-
-      optarg = coil::optarg;
-      optind = coil::optind;
-      optopt = coil::optopt;
-
-      return result;
-    }
+    int operator()();
 
     char* optarg;     //! オプション引数
     int optind;       //! 処理対象引数
